FastAPI/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloud-sql-python-connector==1.4.2
cryptography==41.0.5
distro==1.8.0
ecdsa==0.18.0
exceptiongroup==1.1.3
fastapi==0.104.1
frozenlist==1.4.0
google-auth==2.23.4
greenlet==3.0.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
multidict==6.0.4
mysql==0.0.3
mysql-connector-python==8.2.0
mysqlclient==2.2.0
numpy==1.26.1
openai==1.1.2
packaging==23.2
pandas==2.1.2
passlib==1.7.4
protobuf==4.21.12
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
PyMySQL==1.1.0
pypdf==3.17.0
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.24.0.post1
wget==3.2
yarl==1.9.2
